requirements.txt
natsort==7.1.1
pikepdf==2.5.2
Pillow==8.1.0
